+++ services/analytics-service/app/api/controllers/flows.js
@@ -1,13 +1,12 @@
 const express = require('express');
 const bodyParser = require('body-parser');
-const mongoose = require('mongoose');
-const { reportHealth } = require('../../utils/eventBus');
-const { reportServiceStatus } = require('../../utils/helpers');
+// const mongoose = require('mongoose');
 
 const jsonParser = bodyParser.json();
 const router = express.Router();
 
 router.get('/', jsonParser, async (req, res) => {
+  let response;
   try {
     return res.status(200).send(response);
   } catch (e) {
