+++ services/analytics-service/app/server.js
@@ -6,7 +6,7 @@
 const express = require('express');
 
 const swaggerUi = require('swagger-ui-express');
-const iamMiddleware = require('@openintegrationhub/iam-utils');
+// const iamMiddleware = require('@openintegrationhub/iam-utils');
 const cors = require('cors');
 const config = require('./config/index');
 
@@ -50,13 +50,16 @@ class Server {
 
     // Enables preflight OPTIONS requests
     this.app.options('/', cors());
+
+    // enables CORS
+    this.app.use('/flows', cors(corsOptions));
   }
 
-  async setupMiddleware() {
-    log.info('Setting up middleware');
-
-    return next();
-  }
+  // async setupMiddleware() {
+  //   log.info('Setting up middleware');
+  //
+  //   return next();
+  // }
 
   // This middleware compiles the relevant membership ids generated by the IAM iam middleware and passes them on
   // this.app.use('/storedFunction', async (req, res, next) => {
